refactor: 각 API 목적 통합 및 불필요한 파일 제거와 타입 해당 API 파일 상단으로 이동
