--- src/api/deleteCartItem.ts
@@ -1,8 +1,0 @@
-import { baseAPI } from './baseAPI';
-
-export async function deleteCartItem(cartId: string) {
-  await baseAPI({
-    method: 'DELETE',
-    path: `/cart-items/${cartId}`,
-  });
-}
